pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "python-boilerplate"
version = "0.1.0"
description = "Opinionated Python service boilerplate with uv and mandatory observability."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
  { name = "Nathaniel", email = "hartmannlight@gmail.com" },
]
dependencies = [
  "opentelemetry-exporter-otlp-proto-http>=1.33.0",
  "opentelemetry-sdk>=1.33.0",
  "prometheus-client>=0.22.0",
  "sentry-sdk>=2.26.0",
  "structlog>=25.2.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.11",
]

[project.scripts]
boilerplate = "python_boilerplate.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/python_boilerplate"]

[dependency-groups]
dev = [
  "mypy>=1.15",
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "ruff>=0.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src", "tests"]

[tool.uv]
default-groups = ["dev"]
